Update cache file location to support awscli 1.14+
As described in #1 awscli has change the internal cache file names
and is using a sha1 hash of the serialized arguments used to assume
the role. Since we're always using the same sts call to generate the
session we can recreate the JSON string to derive the hash used in
the cache file name.